A huge amount of fixes, cleanup & documentation.
Added description of the plugin to the README so that people know what
it is. Clarified in the README that taggable objects need to be
persisted. Added documentation covering listing Taggable objects that
match a given tag. Added documentation to encourage people to contribute
issues as well as code. I also added exception throwing in the case
where non-persisted taggable objects are attempted to be interacted
with. I also corrected Taggable tags method to return tags in descending
order of number of references.