Reworked tags(taggable) to return correct values.
The Tagger.tags(taggable) method was return incorrect values previously.
I reworked the query to be correct and added test coverage to verify it.
I am not sure that the forcing of cache breaking is needed and am going
to revert the previous commit changes that forced the cache breaking
until I decide if it is actually necessary.