Tags methods now order ascending alphabetically.
I made this changed because I figured this would be the most common
case. If we find that other cases are necessary I will add them as we
need them via an optional argument to the tags methods. I also renamed
the has_tag? method on taggables to tagged_with? as it is a more correct
method name.