Made solr startup wait dynamically.
I did this because waiting a fixed amount of time didn't seem to be
working very well for different machines on Travis CI and it was making
my local runtime take forever. Therefore, I have it check up to fifteen
times now and as soon as it finds the Solr instance running it stops
checking and runs the tests.