Cleaned up Sunspot ActiveRecord integration.
I did this for two reasons. One, it should solve the tags not being
indexed properly as it normally happens in Rails apps problem. Plus, it
also cleans up the code base. Beyond that it resolves in issue where the
Tagalong library was causing the client Rails apps rake sunspot:reindex
calls to break. Beyond that it allowed my to clean up the spec_helper
and now use sunspot_test which helps by stubbing out the sunspot session
normally and only including it on tests that are tagged with :search =>
true.