Monkey patched SunspotTest to use Solr directly.
The default server method that is provided inside the SunspotTest module
uses Sunspot::Rails::Server to create the Solr server. However, this
doesn't work properly in this gem because this is simply a ruby lib and
not a Rails app. Therefore, I have overwritten the server method to
return an instance of Sunspot::Solr::Server which does NOT depend on
Rails.